add models relations

+++ database/migrations/2020_11_10_103911_create_wallets_table.php
@@ -16,7 +16,7 @@ class CreateWalletsTable extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id')->index();
-            $table->float('balance')->default(0);
+            $table->double('balance')->default(0);
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
